[Finished] Project Version 2/requirements.txt
numpy
pandas
matplotlib
seaborn
Pillow
scikit-learn
scipy
tensorflow
streamlit
